register,login,post

+++ app/Http/Controllers/TasksController.php
@@ -18,7 +18,7 @@ class TasksController extends Controller
         $data = [];
         if (\Auth::check()) {
             $user = \Auth::user();
-            $tasks = $user->feed_microposts()->orderBy('created_at', 'desc')->paginate(10);
+            $tasks = $task->feed_tasks()->orderBy('created_at', 'desc')->paginate(10);
 
             $data = [
                 'user' => $user,
@@ -51,7 +51,7 @@ class TasksController extends Controller
         $this->validate($request, ['status' => 'required|max:10', 'content' => 'required|max:191',]);
         $task = new Task;
         $task->status = $request->status;
-        $task->content = $request->content;
+        $task->content = $request->user()->tasks()->create(['content' => $request->content,]);
         $task->save();
         
         return redirect('/');
@@ -109,8 +109,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+        
+        if(\Auth::id === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
